Adds support for os-auth-strategy=noauth
The noauth strategy will not attempt to look for a token and will not use a
token if one is given. This strategy will also not attempt to find a catalog
and therefore the endpoint must be specified (either in the constructor or
using -os-url).

If the noauth strategy is specified and the endpoint is not it will fail with
a descriptive message.

A small amount of branching logic was added to do_request to only add the
X-Auth-Token when the auth strategy is keystone. There is a blueprint (see:
pluggable-neutronclient-auth) which should implement this logic in a more
elegant way.

A basic test was made to ensure that the auth strategy acted as designed.

Implements: blueprint support-neutronclient-noauth
Change-Id: Id875ce6a4dfd1f932aaab2c25d430ed60e4713bb