Client command extension support
The extension support provided by this change is limited to shell extensions.
The library is not extended by this change. Another change to make extensions
work as library functions will be needed. If desired this change can be
expanded to include that functionality.

Extensions are installed by either placing a conforming module in the
appropriate directory (/v2.0/contrib) or by installing a package into the
environment with a name that ends with '_python_neutronclient_ext'.

A conforming module, at a minimum:
* exposes new commands through a dictionary whose key is the command and value
  is a class that subclasses an extension.ClientExtension* class
* provides classes that subclass extension.ClientExtension* classes

Provides client command extensions through a new class, NeutronClientExtension.
The precedence of command loading are as follows:
* hard coded commands are loaded first
* contribued commands (those in /contrib)
* external commands (installed in the environment) are loaded last

Commands that have the same name will be overwritten by commands that are
loaded later. To greatly change the execution of a command for your particular
extension you only need to override the execute method.

An example extension, while not included, can be seen at:
https://github.com/Cerberus98/quark_python_neutronclient_ext

Change-Id: I5b2fe530c90b5ce1243fc10341d6d434a1ecea7a
Implements: blueprint extensible-neutronclient