Client command extension support
Adds extension support with emphasis on ease of extension creation. Extensions
strongly conform to preexisting neutron commands (/neutron/v2_0/*). A sample
extension has been included (/neutron/v2_0/contrib/fox_sockets.py). As it is
assumed that the sample extension will be packaged with the client, small
changes were required to include it with the unit tests.

Extensions are installed by either placing a conforming module in the
appropriate directory (/neutron/v2.0/contrib) or by installing a package into
the environment with a name that ends with '_python_neutronclient_ext'.
Extension discovery is modeled after nova's module discovery but deviates
strongly beyond that.

A conforming module, at a minimum:
* Will have a class that subclasses NeutronClientExtension to provide the
  requisite version support, paths, and variable names for the client.
  Example: neutronclient.neutron.v2_0.contrib.fox_sockets.FoxInSocket
* Will have at least one class that subclasses from the ClientExtension*
  classes to provide the new functionality to the client
  Example: neutronclient.neutron.v2_0.contrib.fox_sockets.FoxInSocketsList
* ClientExtension* subclasses must have a shell_command class variable if the
  command is to be available to the CLI (shell.py)
  Example: neutronclient.neutron.v2_0.contrib.fox_sockets.FoxInSocketsList

Provides client command extensions through new classes: NeutronClientExtension,
and ClientExtension<Action>.
The precedence of command loading are as follows:
* hard coded commands are loaded first
* contribued commands (those in /contrib)
* external commands (installed in the environment) are loaded last

Commands that have the same name will be overwritten by commands that are
loaded later. To greatly change the execution of a command for your particular
extension you only need to override the execute method.

Currently this extension support is limited to top-level resources. Parent/
child relationships may be added if desired.

Change-Id: I5b2fe530c90b5ce1243fc10341d6d434a1ecea7a
Implements: blueprint extensible-neutronclient